Moved DoctrineToGatewayEventSubscr code to ObjectEntitySubscriber
