+++ src/Subscriber/ObjectEntitySubscriber.php
@@ -3,6 +3,7 @@
 namespace CommonGateway\CoreBundle\Subscriber;
 
 use App\Entity\ObjectEntity;
+use App\Event\ActionEvent;
 use App\Service\ObjectEntityService;
 use CommonGateway\CoreBundle\Message\CacheMessage;
 use CommonGateway\CoreBundle\Service\AuditTrailService;
@@ -12,10 +13,10 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Events;
 use Doctrine\Persistence\Event\LifecycleEventArgs;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Messenger\MessageBusInterface;
 
 /**
  * This subscriber listens for events related to ObjectEntities.
@@ -83,6 +84,20 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
     private SessionInterface $session;
 
     /**
+     * The message bus.
+     *
+     * @var MessageBusInterface $messageBus
+     */
+    private MessageBusInterface $messageBus;
+    
+    /**
+     * Event Dispatcher.
+     *
+     * @var EventDispatcherInterface $eventDispatcher
+     */
+    private EventDispatcherInterface $eventDispatcher;
+    
+    /**
      * The constructor sets al needed variables.
      *
      * @param ObjectEntityService    $objectEntityService   The Object Entity Service.
@@ -92,6 +107,8 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
      * @param CacheService           $cacheService          The cache service.
      * @param AuditTrailService      $auditTrailService     The Audit Trail service.
      * @param SessionInterface       $session               The current session.
+     * @param MessageBusInterface    $messageBus            The messageBus for async messages.
+     * @param EventDispatcherInterface $eventDispatcher     Event Dispatcher.
      */
     public function __construct(
         ObjectEntityService $objectEntityService,
@@ -101,6 +118,8 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
         CacheService $cacheService,
         AuditTrailService $auditTrailService,
         SessionInterface $session,
+        MessageBusInterface $messageBus,
+        EventDispatcherInterface $eventDispatcher
     ) {
         $this->objectEntityService = $objectEntityService;
         $this->entityManager       = $entityManager;
@@ -109,25 +128,102 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
         $this->cacheService        = $cacheService;
         $this->auditTrailService   = $auditTrailService;
         $this->session             = $session;
+        $this->messageBus          = $messageBus;
+        $this->eventDispatcher     = $eventDispatcher;
 
     }//end __construct()
 
     /**
      * Defines the events that the subscriber should subscribe to.
+     * This method can only return the event names.
+     * you cannot define a custom method name to execute when each event triggers.
      *
      * @return array The subscribed events
      */
     public function getSubscribedEvents(): array
     {
         return [
+            Events::prePersist,
+            Events::preUpdate,
             Events::preRemove,
             Events::postLoad,
             Events::postPersist,
             Events::postUpdate,
+            Events::postRemove,
         ];
 
     }//end getSubscribedEvents()
-
+    
+    /**
+     * Handles prePersist Event.
+     *
+     * @param LifecycleEventArgs $args The lifecycle event arguments for this prePersist event.
+     */
+    public function prePersist(LifecycleEventArgs $args): void
+    {
+        $object = $args->getObject();
+        if ($object instanceof ObjectEntity === false) {
+            return;
+        }
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Creating object in database.'
+        // Write the log.
+        $this->logger->info(
+            'Creating object in database',
+            [
+                'object' => $object->getId(),
+                'entity' => $object->getEntity()->getId(),
+            ]
+        );
+        
+        // Throw the event.
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            [
+                'object' => $object,
+                'entity' => ['id' => $object->getEntity()->getId(),'reference' => $object->getEntity()->getReference()]
+            ],
+            'commongateway.object.pre.create'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
+        
+    }//end prePersist()
+    
+    /**
+     * Handles preUpdate Event.
+     *
+     * @param LifecycleEventArgs $args The lifecycle event arguments for this preUpdate event.
+     */
+    public function preUpdate(LifecycleEventArgs $args): void
+    {
+        $object = $args->getObject();
+        if ($object instanceof ObjectEntity === false) {
+            return;
+        }
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Updating object to database.'
+        // Write the log.
+        $this->logger->info(
+            'Updating object to database',
+            [
+                'object' => $object->getId(),
+                'entity' => $object->getEntity()->getId(),
+            ]
+        );
+        
+        // Throw the event.
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            [
+                'object' => $object,
+                'entity' => ['id' => $object->getEntity()->getId(),'reference' => $object->getEntity()->getReference()]
+            ],
+            'commongateway.object.pre.update'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
+        
+    }//end preUpdate()
+    
     /**
      * Handles preRemove Event.
      *
@@ -158,6 +254,28 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
         
         // TODO: old CacheDatabaseSubscriber code: 'Remove objects from the cache after they are removed from the database.'
         $this->cacheService->removeObject($object);
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Deleting object from database.'
+        // Write the log.
+        $this->logger->info(
+            'Deleting object from database',
+            [
+                'object' => $object->getId(),
+                'entity' => $object->getEntity()->getId(),
+            ]
+        );
+        
+        // Throw the event.
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            [
+                'object' => $object,
+                'entity' => ['id' => $object->getEntity()->getId(),'reference' => $object->getEntity()->getReference()]
+            ],
+            'commongateway.object.pre.delete'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
+
     }//end preRemove()
 
     /**
@@ -190,6 +308,23 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
 
             $this->auditTrailService->createAuditTrail($object, $config);
         }
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Read object from database.'
+        // Write the log
+        $this->logger->info(
+            'Read object from database',
+            [
+                'object' => $object->getId(),
+            ]
+        );
+        
+        // Throw the event
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            ['object' => $object],
+            'commongateway.object.post.read'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
 
     }//end postLoad()
 
@@ -245,6 +380,27 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
             // Dispatch event.
             $this->objectEntityService->dispatchEvent('commongateway.action.event', $data, 'commongateway.object.sync');
         }
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Created object in database.'
+        // Write the log.
+        $this->logger->info(
+            'Created object in database',
+            [
+                'object' => $object->getId(),
+                'entity' => $object->getEntity()->getId(),
+            ]
+        );
+        
+        // Throw the event.
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            [
+                'object' => $object,
+                'entity' => ['id' => $object->getEntity()->getId(), 'reference' => $object->getEntity()->getReference()]
+            ],
+            'commongateway.object.post.create'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
 
     }//end postPersist()
 
@@ -289,6 +445,56 @@ class ObjectEntitySubscriber implements EventSubscriberInterface
                 $this->auditTrailService->createAuditTrail($object, $config);
             }
         }//end if
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Updated object in database.'
+        // Write the log.
+        $this->logger->info(
+            'Updated object in database',
+            [
+                'object' => $object->getId(),
+                'entity' => $object->getEntity()->getId(),
+            ]
+        );
+        
+        // Throw the event.
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            [
+                'object' => $object,
+                'entity' => ['id' => $object->getEntity()->getId(), 'reference' => $object->getEntity()->getReference()]
+            ],
+            'commongateway.object.post.update'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
 
     }//end postUpdate()
+    
+    /**
+     * Handles postRemove Event.
+     *
+     * @param LifecycleEventArgs $args The lifecycle event arguments for this postRemove event.
+     */
+    public function postRemove(LifecycleEventArgs $args): void
+    {
+        $object = $args->getObject();
+        if ($object instanceof ObjectEntity === false) {
+            return;
+        }
+        
+        // TODO: old DoctrineToGatewayEventSubscriber code: 'Deleted object from database.'
+        // Write the log.
+        $this->logger->info(
+            'Deleted object from database',
+            []
+        );
+        
+        // Throw the event.
+        $event = new ActionEvent(
+            'commongateway.action.event',
+            [],
+            'commongateway.object.post.delete'
+        );
+        $this->eventDispatcher->dispatch($event, 'commongateway.action.event');
+        
+    }//end postRemove()
 }//end class
